requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aioboto3=12.0.0=pypi_0
aiobotocore=2.7.0=pypi_0
aiohttp=3.8.6=pypi_0
aioitertools=0.11.0=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
blas=1.0=mkl
boto3=1.28.64=pypi_0
botocore=1.31.64=pypi_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.5.7=hbcca054_0
cachetools=5.3.0=pypi_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py39h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cudatoolkit=10.2.89=h713d32c_10
cupy=12.0.0=py39h505621e_1
cycler=0.11.0=pypi_0
deeplake=3.8.6=pypi_0
dill=0.3.7=pypi_0
einops=0.6.1=pypi_0
fastrlock=0.8=py39h5a03fae_2
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.39.4=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fsspec=2023.9.2=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.18.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.0=pypi_0
huggingface-hub=0.17.3=pypi_0
humbug=0.3.2=pypi_0
idna=3.4=py39h06a4308_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py39h06a4308_0
jmespath=1.0.1=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeeplake=0.0.89=pypi_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lmdb=1.4.1=pypi_0
lz4=4.3.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.7.1=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py39h06a4308_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_1
numpy=1.22.0=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=23.1=pypi_0
pandas=2.0.1=pypi_0
pathos=0.3.1=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
pox=0.3.3=pypi_0
ppft=1.7.6.7=pypi_0
prettytable=3.9.0=pypi_0
protobuf=4.23.0=pypi_0
pyarrow=13.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.5.1=pypi_0
pydantic-core=2.14.3=pypi_0
pyjwt=2.8.0=pypi_0
pyopenssl=23.0.0=py39h06a4308_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python_abi=3.9=2_cp39
pytorch=2.0.1=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.29.0=py39h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.7.0=pypi_0
safetensors=0.4.0=pypi_0
scikit-learn=1.3.1=pypi_0
scipy=1.10.1=pypi_0
setuptools=66.0.0=py39h06a4308_0
six=1.16.0=pypi_0
spikingjelly=0.0.0.0.14=pypi_0
sqlite=3.41.2=h5eee18b_0
style=1.1.0=pypi_0
sympy=1.11.1=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.14.1=pypi_0
torchaudio=2.0.2=py39_cu118
torchtoolbox=0.1.8.2=pypi_0
torchtriton=2.0.0=py39
torchvision=0.15.2=py39_cu118
tqdm=4.65.0=pypi_0
transformers=4.34.0=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
update=0.0.1=pypi_0
urllib3=1.26.15=py39h06a4308_0
wcwidth=0.2.8=pypi_0
werkzeug=2.3.4=pypi_0
wheel=0.38.4=py39h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
